requirements.txt
# ─── Core Framework ────────────────────────────────────────────────────────────
streamlit>=1.32.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# ─── Vector Database ──────────────────────────────────────────────────────────
chromadb>=0.5.0

# ─── PDF Processing ───────────────────────────────────────────────────────────
pypdf>=4.0.0

# ─── Embeddings ───────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ─── Environment Variables ─────────────────────────────────────────────────────
python-dotenv>=1.0.0
